Fix some static analysis warnings

+++ Windows/W32Util/Misc.cpp
@@ -55,9 +55,9 @@ namespace W32Util
 		}
 		float f = (float)size + ((float)frac / 1024.0f);
 		if (s==0)
-			sprintf(out,"%d B",size);
+			sprintf(out, "%d B", (int)size);
 		else
-			sprintf(out,"%3.1f %s",f,sizes[s]);
+			sprintf(out, "%3.1f %s", f, sizes[s]);
 	}
 
 	BOOL CopyTextToClipboard(HWND hwnd, const char *text) {
